+++ src/keybind/keybind.cpp
@@ -1,221 +1,20 @@
 #include "keybind.hpp"
-#include "../audio.hpp"
-#include "../config/config.hpp"
 #include "../enums.hpp"
 #include "../gdata.hpp"
-#include "../lib/profiler.hpp"
+#include "../log.hpp"
 #include "../variables.hpp"
-#include "keybind_deserializer.hpp"
-#include "keybind_serializer.hpp"
+#include "keybind_manager.hpp"
 
 using namespace std::literals::string_literals;
 
 namespace elona
 {
 
-namespace
-{
-
-// The last held key has to be global across all input contexts, because a key
-// can generate two different actions in different contexts. If key delay
-// depended on the last action generated by a keypress, when the context is
-// changed by entering a menu, there would be no key delay and the key would be
-// counted as pressed twice.
-
-snail::Key last_held_key = snail::Key::none;
-int last_held_key_frames = 0;
-
-} // namespace
-
-// This map holds the action categories available in an input context.
-// Categories at the top will override categories further down. This is to
-// support the behavior of toggling wizard mode.
-// TODO: It still needs to support ignoring key conflicts for togglable
-// categories.
-
-// clang-format off
-static std::map<InputContextType, std::vector<ActionCategory>> input_context_types =
-{
-    {InputContextType::menu, {ActionCategory::shortcut,
-                              ActionCategory::menu,
-                              ActionCategory::selection,
-                              ActionCategory::default_}},
-
-    {InputContextType::game, {ActionCategory::wizard,
-                              ActionCategory::game,
-                              ActionCategory::shortcut,
-                              ActionCategory::default_}}
-};
-// clang-format on
-
-
-KeybindManager& KeybindManager::instance()
-{
-    static KeybindManager the_instance;
-    return the_instance;
-}
-
-KeybindManager::GroupedMapType KeybindManager::create_category_to_action_list()
-{
-    GroupedMapType result;
-
-    for (const auto& pair : keybind::actions)
-    {
-        const auto& action_id = pair.first;
-        const auto& action = pair.second;
-        result.emplace(action.category, action_id);
-    }
-
-    return result;
-}
-
 void initialize_keybindings()
 {
     auto& keybind_manager = KeybindManager::instance();
     initialize_keybind_actions(keybind::actions);
     keybind_manager.load();
-}
-
-void KeybindManager::save()
-{
-    auto path = filesystem::dir::exe() / u8"keybindings.hcl";
-    std::ofstream file{path.native(), std::ios::binary};
-    if (!file)
-    {
-        throw std::runtime_error{
-            u8"Failed to open: "s
-            + filesystem::make_preferred_path_in_utf8(path)};
-    }
-
-    KeybindSerializer(*this).save(file);
-}
-
-void KeybindManager::load()
-{
-    clear();
-    initialize_known_actions(keybind::actions);
-    load_permanent_bindings(keybind::actions);
-
-    auto path = filesystem::dir::exe() / u8"keybindings.hcl";
-    std::ifstream file{path.native(), std::ios::binary};
-
-    if (file)
-    {
-        KeybindDeserializer(*this).load(file);
-    }
-    else
-    {
-        load_default_bindings(keybind::actions);
-        save();
-    }
-}
-
-
-void KeybindManager::initialize_known_actions(const ActionMap& actions)
-{
-    for (const auto& pair : actions)
-    {
-        const auto& action_id = pair.first;
-
-        register_binding(action_id);
-    }
-}
-
-void KeybindManager::load_permanent_bindings(const ActionMap& actions)
-{
-    for (const auto& pair : actions)
-    {
-        const auto& action_id = pair.first;
-        const auto& action = pair.second;
-
-        auto& the_binding = binding(action_id);
-        for (const auto& keybind : action.default_keybinds)
-        {
-            if (!keybind_is_bindable_key(keybind.main))
-            {
-                // Permanently bind keys that shouldn't be unbound (escape,
-                // enter, and directional keys)
-                the_binding.permanent = keybind;
-            }
-        }
-    }
-}
-
-void KeybindManager::load_default_bindings(const ActionMap& actions)
-{
-    for (const auto& pair : actions)
-    {
-        const auto& action_id = pair.first;
-        const auto& action = pair.second;
-
-        auto& the_binding = binding(action_id);
-        for (const auto& keybind : action.default_keybinds)
-        {
-            if (keybind_is_bindable_key(keybind.main))
-            {
-                if (the_binding.primary.empty())
-                {
-                    the_binding.primary = keybind;
-                }
-                else if (the_binding.alternate.empty())
-                {
-                    the_binding.alternate = keybind;
-                }
-                else
-                {
-                    assert(false);
-                }
-            }
-        }
-    }
-}
-
-std::vector<std::string> KeybindManager::find_conflicts(
-    const std::string& action_id,
-    const Keybind& keybind)
-{
-    auto action_category = keybind::actions.at(action_id).category;
-    std::unordered_set<ActionCategory> can_conflict_with;
-    std::vector<std::string> conflicts;
-
-    // Iterate through possible combinations of action categories to see which
-    // one the action is a part of.
-    for (const auto& pair : input_context_types)
-    {
-        const auto& categories = pair.second;
-        bool is_in_context =
-            std::find(categories.begin(), categories.end(), action_category)
-            != categories.end();
-
-        // If the action can be returned from input in this category, the
-        // keybinding for it cannot confict with other actions in other
-        // categories for the same input context type.
-        if (is_in_context)
-        {
-            can_conflict_with.insert(categories.begin(), categories.end());
-        }
-    }
-
-    for (const auto& pair : keybind::actions)
-    {
-        const auto& action_id = pair.first;
-        const auto& action = pair.second;
-        bool action_could_conflict =
-            can_conflict_with.find(action.category) != can_conflict_with.end();
-
-        if (action_could_conflict)
-        {
-            const auto& action_binding = binding(action_id);
-            if (action_binding.primary == keybind
-                || action_binding.alternate == keybind
-                /* || (keybind.is_joystick() && keybind.joystick == keybind.main) */)
-            {
-                conflicts.emplace_back(action_id);
-            }
-        }
-    }
-
-    return conflicts;
 }
 
 std::string Keybind::to_string() const
@@ -296,529 +95,6 @@ optional<Keybind> Keybind::from_string(std::string str)
     return none;
 }
 
-bool InputContext::_matches(
-    const std::string& action_id,
-    snail::Key key,
-    snail::ModKey modifiers)
-{
-    if (_available_actions.find(action_id) == _available_actions.end())
-    {
-        return false;
-    }
-
-    const auto& binding = KeybindManager::instance().binding(action_id);
-
-    if (binding.primary.main == key && binding.primary.modifiers == modifiers)
-    {
-        return true;
-    }
-    if (binding.alternate.main == key
-        && binding.alternate.modifiers == modifiers)
-    {
-        return true;
-    }
-    if (binding.permanent.main == key
-        && binding.permanent.modifiers == modifiers)
-    {
-        return true;
-    }
-    if (binding.joystick == key)
-    {
-        return true;
-    }
-
-    return false;
-}
-
-optional<std::string> InputContext::_action_for_key(const Keybind& keybind)
-{
-    for (const auto& action_id : _available_actions_sorted)
-    {
-        const auto& binding = KeybindManager::instance().binding(action_id);
-        bool excluded =
-            _excluded_categories.find(keybind::actions.at(action_id).category)
-            != _excluded_categories.end();
-
-        if (!excluded && binding.matches(keybind))
-        {
-            return action_id;
-        }
-    }
-
-    return none;
-}
-
-optional<std::string> InputContext::_check_movement_action(
-    const std::set<snail::Key>& keys,
-    snail::ModKey modifiers)
-{
-    StickKey input = StickKey::none;
-
-    // Movement keys have to ignore Shift or Alt. They could be used with them
-    // in a hardcoded manner for running or diagonal restriction.
-    // TODO: At least make this clear to players somehow.
-    snail::ModKey key_modifiers =
-        modifiers & ~(snail::ModKey::shift | snail::ModKey::alt);
-
-    for (const auto& key : keys)
-    {
-        // NOTE: Escape acts as both "summon the quit menu" at the main game
-        // loop and "cancel", but actual checks for escape being pressed are
-        // relatively few. Setting the global key_escape = true is to avoid
-        // having to do (action == "cancel" || action == "escape") everywhere.
-        if (_matches("escape", key, key_modifiers))
-        {
-            bool just_pressed = keywait == 0;
-            keywait = 1;
-            key_escape = true;
-
-            if (just_pressed)
-            {
-                return "cancel"s;
-            }
-            else
-            {
-                return ""s;
-            }
-        }
-
-        if (_matches("north", key, key_modifiers))
-        {
-            input |= StickKey::up;
-        }
-        else if (_matches("south", key, key_modifiers))
-        {
-            input |= StickKey::down;
-        }
-        else if (_matches("east", key, key_modifiers))
-        {
-            input |= StickKey::left;
-        }
-        else if (_matches("west", key, key_modifiers))
-        {
-            input |= StickKey::right;
-        }
-        else if (_matches("northwest", key, key_modifiers))
-        {
-            input = StickKey::up | StickKey::left;
-        }
-        else if (_matches("northeast", key, key_modifiers))
-        {
-            input = StickKey::up | StickKey::right;
-        }
-        else if (_matches("southwest", key, key_modifiers))
-        {
-            input = StickKey::down | StickKey::left;
-        }
-        else if (_matches("southeast", key, key_modifiers))
-        {
-            input = StickKey::down | StickKey::right;
-        }
-        else if (!is_modifier(key))
-        {
-            // Encountered non-movement key, prioritize it over movement.
-            return none;
-        }
-    }
-
-    if ((modifiers & snail::ModKey::shift) == snail::ModKey::shift)
-    {
-        // Has to be modified globally, since scroll speed is determined by
-        // keybd_wait. See @ref ui_scroll_screen()
-        keybd_wait = 100000;
-        if (keywait == 0)
-        {
-            keywait = 1;
-            return "cancel"s;
-        }
-    }
-    else
-    {
-        keywait = 0;
-    }
-
-
-    if ((modifiers & snail::ModKey::alt) != snail::ModKey::alt)
-    {
-        if (input == StickKey::left)
-        {
-            return "west"s;
-        }
-        if (input == StickKey::up)
-        {
-            return "north"s;
-        }
-        if (input == StickKey::right)
-        {
-            return "east"s;
-        }
-        if (input == StickKey::down)
-        {
-            return "south"s;
-        }
-    }
-    if (input == (StickKey::up | StickKey::left))
-    {
-        return "northwest"s;
-    }
-    if (input == (StickKey::up | StickKey::right))
-    {
-        return "northeast"s;
-    }
-    if (input == (StickKey::down | StickKey::left))
-    {
-        return "southwest"s;
-    }
-    if (input == (StickKey::down | StickKey::right))
-    {
-        return "southeast"s;
-    }
-
-    return none;
-}
-
-bool InputContext::_is_nonmovement_key(const snail::Key& k)
-{
-    return keybind_is_bindable_key(k) || k == snail::Key::enter;
-}
-
-optional<Keybind> InputContext::_check_normal_action()
-{
-    const auto& keys = snail::Input::instance().pressed_keys();
-    auto modifiers = snail::Input::instance().modifiers();
-
-    // Pick the first nonmovement key out of the ones that were held.
-    // The only actions for which holding multiple keys makes sense is for the
-    // movement keys.
-    const auto it =
-        std::find_if(keys.begin(), keys.end(), [this](const snail::Key& k) {
-            return _is_nonmovement_key(k);
-        });
-
-    if (it != keys.end())
-    {
-        return Keybind(*it, modifiers);
-    }
-
-    return none;
-}
-
-InputContext InputContext::create(InputContextType type)
-{
-    InputContext result;
-
-    const auto& categories = input_context_types.at(type);
-    for (const auto& category : categories)
-    {
-        result._add_actions_from_category(category);
-    }
-
-    return result;
-}
-
-void InputContext::_add_actions_from_category(ActionCategory category)
-{
-    for (const auto& pair : keybind::actions)
-    {
-        const auto& action_id = pair.first;
-        const auto& action = pair.second;
-
-        if (action.category == category)
-        {
-            _available_actions.insert(action_id);
-            _available_actions_sorted.push_back(action_id);
-        }
-    }
-}
-
-std::string InputContext::_delay_movement_action(
-    const std::string& action,
-    snail::ModKey modifiers,
-    KeyWaitDelay delay_type)
-{
-    if (keybd_wait >= 100000)
-    {
-        if ((modifiers & snail::ModKey::shift) != snail::ModKey::shift)
-        {
-            keybd_wait = 1000;
-        }
-    }
-
-    if (delay_type == KeyWaitDelay::none)
-    {
-        return action;
-    }
-
-    if (delay_type == KeyWaitDelay::walk_run)
-    {
-        if (keybd_attacking != 0)
-        {
-            if (keybd_wait % Config::instance().attackwait != 0)
-            {
-                return ""s;
-            }
-        }
-        else if (Config::instance().scroll == 0)
-        {
-            if (keybd_wait
-                < Config::instance().walkwait * Config::instance().startrun)
-            {
-                if (keybd_wait % Config::instance().walkwait != 0)
-                {
-                    return ""s;
-                }
-            }
-            else
-            {
-                running = 1;
-                if (keybd_wait < 100000)
-                {
-                    if (keybd_wait % Config::instance().runwait != 0)
-                    {
-                        return ""s;
-                    }
-                }
-            }
-        }
-        // else if (input == StickKey::none)
-        else if (action == "wait"s)
-        {
-            if (keybd_wait < 20)
-            {
-                if (keybd_wait != 0)
-                {
-                    return ""s;
-                }
-            }
-        }
-        else if (keybd_wait > Config::instance().startrun)
-        {
-            if (Config::instance().runscroll == 0)
-            {
-                if (keybd_wait % Config::instance().runwait != 0)
-                {
-                    return ""s;
-                }
-            }
-            running = 1;
-        }
-    }
-    else if (
-        keybd_wait
-        < Config::instance().select_fast_start * Config::instance().select_wait)
-    {
-        if (keybd_wait % Config::instance().select_wait != 0)
-        {
-            return ""s;
-        }
-    }
-    else if (keybd_wait < 1000)
-    {
-        if (keybd_wait % Config::instance().select_fast_wait != 0)
-        {
-            return ""s;
-        }
-    }
-
-    return action;
-}
-
-static bool
-_is_keypress_delayed(int held_frames, int keywait, int initial_keywait)
-{
-    if (held_frames < initial_keywait)
-    {
-        if (held_frames == 0)
-        {
-            return false;
-        }
-    }
-    else if (held_frames % keywait == 0)
-    {
-        return false;
-    }
-
-    return true;
-}
-
-bool InputContext::_delay_normal_action(const Keybind& keybind)
-{
-    if (last_held_key != keybind.main)
-    {
-        last_held_key_frames = 0;
-    }
-
-    bool delayed = _is_keypress_delayed(last_held_key_frames, 1, 20);
-
-    last_held_key_frames++;
-
-    if (last_held_key != keybind.main)
-    {
-        last_held_key = keybind.main;
-    }
-
-    if (delayed)
-    {
-        return true;
-    }
-
-    return false;
-} // namespace elona
-
-std::string InputContext::check_for_command(KeyWaitDelay delay_type)
-{
-    key_escape = false;
-
-    const auto& keys = snail::Input::instance().pressed_keys();
-    auto modifiers = snail::Input::instance().modifiers();
-
-    if (/* !shortcut && */ keyhalt != 0)
-    {
-        if (keys.size() > 0)
-        {
-            keybd_wait = 0;
-            return "";
-        }
-        else
-        {
-            keyhalt = 0;
-        }
-    }
-
-    for (const auto& key : keys)
-    {
-        if (auto n = keybind_key_name(key))
-        {
-        }
-    }
-
-    if (const auto action = _check_movement_action(keys, modifiers))
-    {
-        last_held_key = snail::Key::none;
-        last_held_key_frames = 0;
-
-        // Movement keys have special key delay behavior, so handle them.
-        auto result = _delay_movement_action(*action, modifiers, delay_type);
-        ++keybd_wait;
-
-        return result;
-    }
-    else
-    {
-        keybd_wait = 0;
-        keybd_attacking = 0;
-        running = 0;
-    }
-
-    if (const auto keybind = _check_normal_action())
-    {
-        if (const auto action = _action_for_key(*keybind))
-        {
-            if (!_delay_normal_action(*keybind))
-            {
-                return *action;
-            }
-        }
-        else
-        {
-            last_held_key = snail::Key::none;
-            last_held_key_frames = 0;
-        }
-    }
-    else
-    {
-        last_held_key = snail::Key::none;
-        last_held_key_frames = 0;
-    }
-
-    return ""s;
-}
-
-std::string InputContext::check_for_command_with_list(int& list_index)
-{
-    auto action = check_for_command(KeyWaitDelay::always);
-
-    if (action == "north"s)
-    {
-        snd(5);
-        --cs;
-        if (cs < 0)
-        {
-            cs = keyrange - 1;
-            if (cs < 0)
-            {
-                cs = 0;
-            }
-        }
-    }
-    if (action == "south"s)
-    {
-        snd(5);
-        ++cs;
-        if (cs >= keyrange)
-        {
-            cs = 0;
-        }
-    }
-    if (action == "west"s)
-    {
-        action = "previous_page"s;
-    }
-    if (action == "east"s)
-    {
-        action = "next_page"s;
-    }
-    if (cs >= keyrange)
-    {
-        cs_bk = -1;
-        cs = keyrange - 1;
-        if (cs < 0)
-        {
-            cs = 0;
-        }
-    }
-    if (action == "enter"s)
-    {
-        list_index = cs;
-    }
-    else if (keybind_action_has_category(action, ActionCategory::selection))
-    {
-        list_index = keybind_index_number(action);
-    }
-    else
-    {
-        list_index = -1;
-    }
-    return action;
-}
-
-void InputContext::reset()
-{
-    snail::Input::instance().clear_pressed_keys_and_modifiers();
-    key_escape = false;
-    last_held_key = snail::Key::none;
-    last_held_key_frames = 0;
-}
-
-InputContext& InputContext::instance()
-{
-    static optional<InputContext> the_input_context;
-    if (!the_input_context)
-    {
-        the_input_context = create(InputContextType::game);
-    }
-    return *the_input_context;
-}
-
-InputContext& InputContext::for_menu()
-{
-    static optional<InputContext> the_input_context;
-    if (!the_input_context)
-    {
-        the_input_context = create(InputContextType::menu);
-    }
-    return *the_input_context;
-}
 
 bool keybind_is_joystick_key(snail::Key key)
 {
@@ -846,7 +122,6 @@ bool keybind_action_has_category(
 
     return keybind::actions.at(action_id).category == category;
 }
-
 
 void draw_init_key_select_buffer();
 
@@ -880,100 +155,14 @@ void keybind_regenerate_key_select()
     draw_init_key_select_buffer();
 }
 
-static std::string _binding_name(
-    KeybindManager& keybind_manager,
-    const std::string& action_id)
-{
-    std::string result = "";
-    auto& binding = keybind_manager.binding(action_id);
-    if (!binding.primary.empty())
-    {
-        result = binding.primary.to_string();
-    }
-    else if (!binding.alternate.empty())
-    {
-        result = binding.primary.to_string();
-    }
-    else if (binding.joystick != snail::Key::none)
-    {
-        if (auto name_opt = keybind_key_name(binding.joystick))
-        {
-            result = *name_opt;
-        }
-    }
-
-    strutil::try_remove_prefix(result, "Keypad ");
-    return result;
-}
-
-// TODO: delete this.
-void keybind_regenerate_key_names()
-{
-    auto& km = KeybindManager::instance();
-
-    // clang-format off
-    key_north     = _binding_name(km, "north");
-    key_south     = _binding_name(km, "south");
-    key_west      = _binding_name(km, "west");
-    key_east      = _binding_name(km, "east");
-    key_northwest = _binding_name(km, "northwest");
-    key_northeast = _binding_name(km, "northeast");
-    key_southwest = _binding_name(km, "southwest");
-    key_southeast = _binding_name(km, "southeast");
-    key_wait      = _binding_name(km, "wait");
-    key_cancel    = _binding_name(km, "cancel");
-    key_pageup    = _binding_name(km, "next_page");
-    key_pagedown  = _binding_name(km, "previous_page");
-    key_mode      = _binding_name(km, "switch_mode");
-    key_mode2     = _binding_name(km, "switch_mode_2");
-    key_quick     = _binding_name(km, "quick_menu");
-    key_zap       = _binding_name(km, "zap");
-    key_inventory = _binding_name(km, "inventory");
-    key_quickinv  = _binding_name(km, "quick_inventory");
-    key_get       = _binding_name(km, "get");
-    key_drop      = _binding_name(km, "drop");
-    key_charainfo = _binding_name(km, "chara_info");
-    key_enter     = _binding_name(km, "enter");
-    key_eat       = _binding_name(km, "eat");
-    key_wear      = _binding_name(km, "wear");
-    key_cast      = _binding_name(km, "cast");
-    key_drink     = _binding_name(km, "drink");
-    key_read      = _binding_name(km, "read");
-    key_fire      = _binding_name(km, "fire");
-    key_godown    = _binding_name(km, "go_down");
-    key_goup      = _binding_name(km, "go_up");
-    key_save      = _binding_name(km, "save");
-    key_search    = _binding_name(km, "search");
-    key_interact  = _binding_name(km, "interact");
-    key_identify  = _binding_name(km, "identify");
-    key_skill     = _binding_name(km, "skill");
-    key_close     = _binding_name(km, "close");
-    key_rest      = _binding_name(km, "rest");
-    key_target    = _binding_name(km, "target");
-    key_dig       = _binding_name(km, "dig");
-    key_use       = _binding_name(km, "use");
-    key_bash      = _binding_name(km, "bash");
-    key_open      = _binding_name(km, "open");
-    key_dip       = _binding_name(km, "dip");
-    key_pray      = _binding_name(km, "pray");
-    key_offer     = _binding_name(km, "offer");
-    key_journal   = _binding_name(km, "journal");
-    key_material  = _binding_name(km, "material");
-    key_trait     = _binding_name(km, "trait");
-    key_look      = _binding_name(km, "look");
-    key_give      = _binding_name(km, "give");
-    key_throw     = _binding_name(km, "throw");
-    key_ammo      = _binding_name(km, "ammo");
-    key_autodig   = _binding_name(km, "autodig");
-    key_quicksave = _binding_name(km, "quicksave");
-    key_quickload = _binding_name(km, "quickload");
-    key_help      = _binding_name(km, "help");
-    key_msglog    = _binding_name(km, "message_log");
-    // clang-format on
-}
-
 int keybind_index_number(const std::string& action_id)
 {
+    if (!(keybind_action_has_category(action_id, ActionCategory::selection)
+          || keybind_action_has_category(action_id, ActionCategory::shortcut)))
+    {
+        return 0;
+    }
+
     auto underscore_pos = action_id.find("_");
     assert(underscore_pos != std::string::npos);
