+++ src/keybind/keybind_deserializer.cpp
@@ -3,6 +3,7 @@
 #include <string>
 #include "../hcl.hpp"
 #include "keybind.hpp"
+#include "keybind_manager.hpp"
 
 namespace elona
 {
